Issue 20: write uncompressed bytes if uncompressed size==compressed size
The LZO specification says that we should write the uncompressed bytes
rather than the compressed bytes if the compressed buffer is actually
larger than the uncompresesd buffer.

To conform to the standard, this means we have to write the uncompressed
bytes also when they have exactly the same size as the compressed bytes.