Added the boolean option "deprecated.lzo.text.input.format.ignore.non.lzo.extensions", which defaults to true. The option is to be used with the com.hadoop.mapred.DeprecatedLzoTextInputFormat input format class.
When true, it causes all files that don't end in ".lzo" to be silently dropped from
the input set.

When false, it will keep files that don't end in ".lzo", and will process them
with TextInputFormat (however, files that end in ".lzo.index" will still be ignored).
This makes it possible to process a mix of LZO and non-LZ) files with a single
streaming MR job.

Also bumped the version number to 0.4.14.