1) Added the boolean option "lzo.text.input.format.ignore.nonlzo", (default is true). The option is to be used with the DeprecatedLzoTextInputFormat and LzoTextInputFormat input format classes.
When true, it causes all files that don't end in ".lzo" to be silently dropped
from the input set.

When false, it will keep files that don't end in ".lzo", and will process them
with TextInputFormat (however, files that end in ".lzo.index" will still be
ignored). This makes it possible to process a mix of LZO and non-LZO files
with a single MR job, which in turn makes it much easier to perform an online
upgrade to LZO compression in a production system without incurring downtime.

It also makes it possible to reprocess ranges of log files that span the
pre-LZO / post-LZO boundary in a single MR job.

2) Added unit test for the above feature to TestLzoTextInputFormat.

3) Added a public LzopCodec.DEFAULT_LZO_EXTENSION constant.