Added the LzoStreamingInputFormat and LzoStreamingLineRecordReader classes.
These classes are more appropriate than DeprecatedLzoTextInputFormat /
DeprecatedLzoLineRecordReader for use with the hadoop-streaming jar, since
they have the same behavior as the default streaming input format:

- input is broken into lines using any of '\n', '\r', or '\r\n'.
- line contents up to the first '\t' character are treated as the key
- the rest of the line is treated as the value

In contrast, the DeprecatedLzoTextInputFormat treats the file offset as the
key and the entire line as the value. This resulted in weird behavior when
using the DeprecatedLzoTextInputFormat with a streaming MR job. For example,
when using -mapper 'cat' and no reducer (which should produce an output
file that's identical to the input file), this input

key1	    value1
key2	    value2
key3	    value3

Produced this output:

0	 key1 value1
95	 key2 value2
95	 key3 value3

which is clearly wrong. Using LzoStreamingInputFormat produces the expected
output (same as input).